accessibility/visionguide/requirements.txt

fastapi
uvicorn
pillow
google-generativeai
python-multipart
python-dotenv
